modified

+++ partition_opt.py
@@ -397,8 +397,8 @@ def get_score(y_true, y_pred, mode = MODE):
     if len(y_true.shape)==1:
       # y_true = tf.one_hot(y_true, NUM_CLASS)
       # y_pred = tf.one_hot(y_pred, NUM_CLASS)
-      y_true = np.eye(NUM_CLASS)[y_true].astype(int)
-      y_pred = np.eye(NUM_CLASS)[y_pred].astype(int)
+      y_true = np.eye(NUM_CLASS)[y_true.astype(int)].astype(int)
+      y_pred = np.eye(NUM_CLASS)[y_pred.astype(int)].astype(int)
     # else:
     # #this is to make coding consistent for later parts of the function (where tf functions are used)
     #   y_true = tf.convert_to_tensor(y_true)
